Fix dialyzer warning
While using this in another project, we are initializing a payloader and
passing the generated reference to the various other functions. Dialyzer
is currently complaining because `Membrane.Ogg.Payloader`'s spec
currently marks its return as `{:ok, binary}` which is causing a spec
mismatch with the other functions which do require a reference.